Performance changes.
Use hash tables for mapping action ids to groups. Resolve group names to
group ids only once when the application reads the configuration. Keep
the group ids sorted and search them using binary search.

Since the group ids are resolved when the service starts up, it is
theoretically possible that the group ids no longer matches to the group
names, if the group table has been changed so that the pre-existing
group ids are mapped to new group names.

Signed-off-by: Ismo Puustinen <ismo.puustinen@intel.com>